backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
openai==1.37.0
PyMuPDF==1.24.7
python-docx==1.1.2
openpyxl==3.1.5
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.2
python-multipart==0.0.9
requests==2.32.3
